requirements.txt
Flask==0.9
Jinja2==2.6
Werkzeug==0.8.3
argparse==1.2.1
distribute==0.6.24
PyCrypto==2.6.1
pyOpenSSL==0.14
wsgiref==0.1.2
